+++ src/components/board/BoardCategorySlider.jsx
@@ -1,29 +1,14 @@
 import React, { useState } from "react";
+import Categories from "../elem/WholeCategories";
 import { useRef } from "react";
-import * as St from "./BoardCategorySlider.style"
+import * as St from "./BoardCategorySlider.style";
 
 const BoardCategorySlider = ({ parentFunction }) => {
   const [prevCategory, setPrevCategory] = useState("전체");
   const [category, setCategory] = useState("전체");
   const [page, setPage] = useState(1);
 
-  const categories = [
-    "전체",
-    "패션/뷰티",
-    "맛집/요리/음식",
-    "경제/재테크",
-    "썸/연애",
-    "취미/운동",
-    "스포츠",
-    "여행",
-    "결혼",
-    "게임",
-    "반려동물",
-    "가족",
-    "취업/자격증",
-    "일상",
-    "기타",
-  ];
+  const categories = Categories;
 
   const handleCategoryClick = (category) => {
     setPrevCategory(prevCategory);
@@ -52,9 +37,6 @@ const BoardCategorySlider = ({ parentFunction }) => {
       return setCurrentPosition((prev) => prev - 400);
     }
   };
-
-  //   버튼 css 적용
-  //   넘치는 거 지우기
 
   return (
     <>
