+++ src/components/elem/WholeCategories.jsx
@@ -1,0 +1,19 @@
+const Categories = [
+  "전체",
+  "패션/뷰티",
+  "맛집/요리/음식",
+  "경제/재테크",
+  "썸/연애",
+  "취미/운동",
+  "스포츠",
+  "여행",
+  "결혼",
+  "게임",
+  "반려동물",
+  "가족",
+  "취업/자격증",
+  "일상",
+  "기타",
+];
+
+export default Categories;
