+++ src/components/header/BoardHeader.style.js
@@ -1,0 +1,57 @@
+import styled from "styled-components";
+
+const Row = styled.div`
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+`;
+
+const Header = styled.div`
+  background-color: white;
+  position: fixed;
+  top: 0;
+  padding-top: 0.75rem;
+  width: 100%;
+  max-width: 400px;
+  font-size: 1.25rem;
+  color: rgb(70, 70, 70);
+`;
+
+const Wrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  font-weight: bold;
+  justify-content: space-evenly;
+`;
+
+const MainCategory = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  padding-bottom: 0.75rem;
+  text-decoration: none;
+
+  color: ${(props) =>
+    props.isActive ? "rgb(70, 70, 70)" : "rgb(180, 180, 180)"};
+  border-bottom: ${(props) =>
+    props.isActive
+      ? "0.2rem solid rgb(70, 70, 70)"
+      : "0.1rem solid rgb(180, 180, 180)"};
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  &:active {
+    color: rgb(70, 70, 70);
+    border-bottom: 0.2rem solid rgb(70, 70, 70);
+  }
+`;
+
+const Title = styled.div`
+  padding-bottom: 2rem;
+  display: flex;
+  justify-content: center;
+`;
+
+export { Header, Wrap, Title, Row, MainCategory };
