+++ src/components/hook/scroll/InfiniteScroll.jsx
@@ -1,0 +1,30 @@
+import { useEffect } from "react";
+
+const InfiniteScroll = ({ postCardContRef, setPage }) => {
+  useEffect(() => {
+    const handleScroll = () => {
+      const postCardCont = postCardContRef.current;
+
+      if (!postCardCont) {
+        return;
+      }
+
+      const isBottom =
+        postCardCont.scrollHeight - postCardCont.scrollTop ===
+        postCardCont.clientHeight;
+
+      if (isBottom) {
+        setPage((prevPage) => prevPage + 1);
+      }
+    };
+
+    postCardContRef.current?.addEventListener("scroll", handleScroll);
+    return () => {
+      postCardContRef.current?.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+
+  return;
+};
+
+export default InfiniteScroll;
