+++ src/components/post/HotPostCard.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import * as St from "./HotPostCard.style"
+import Categories from "../elem/Categories";
+import Icons from "../elem/Icons";
+import * as St from "./HotPostCard.style";
 
 const HotPostCard = ({
   mainCategory,
@@ -14,49 +16,14 @@ const HotPostCard = ({
 }) => {
   const navigate = useNavigate();
 
-  console.log("likesCount=========================##", likesCount);
-
-  //* =========== Cat. Label ===============
-  const categories = [
-    "패션/뷰티",
-    "맛집/요리/음식",
-    "경제/재테크",
-    "썸/연애",
-    "취미/운동",
-    "스포츠",
-    "여행",
-    "결혼",
-    "게임",
-    "반려동물",
-    "가족",
-    "취업/자격증",
-    "일상",
-    "기타",
-  ];
-
-  const icons = [
-    "👗",
-    "🍱",
-    "💰",
-    "💘",
-    "🎨",
-    "⚽️",
-    "✈️",
-    "👩‍❤️‍👨",
-    "🎮",
-    "🐶",
-    "👨‍👩‍👧‍👦",
-    "💼",
-    "💬",
-    "🎸",
-  ];
+  const categories = Categories;
+  const icons = Icons;
 
   const categoryIconsMap = categories.reduce((acc, cur, idx) => {
     acc[cur] = icons[idx];
     return acc;
   }, {});
 
-  //* =========== Cat. Label ===============
   return (
     <St.PostCardWrap
       mainCategory={mainCategory}
@@ -93,9 +60,7 @@ const HotPostCard = ({
               {/* <St.PostCardInfoContentImg src={like} /> */}
               <St.IconLike />
             </St.InfoContentIcon>
-            <St.InfoContentCount>
-              {likesCount}
-            </St.InfoContentCount>
+            <St.InfoContentCount>{likesCount}</St.InfoContentCount>
           </St.InfoContent>
           {/* 조회수 */}
           <St.InfoContent>
@@ -103,9 +68,7 @@ const HotPostCard = ({
               {/* <StPostCardInfoContentImg src={view} /> */}
               <St.IconView />
             </St.InfoContentIcon>
-            <St.InfoContentCount>
-              {viewCount}
-            </St.InfoContentCount>
+            <St.InfoContentCount>{viewCount}</St.InfoContentCount>
           </St.InfoContent>
           {/* 댓글 버튼, 개수 */}
           <St.InfoContent>
@@ -113,9 +76,7 @@ const HotPostCard = ({
               {/* <StPostCardInfoContentImg src={comment} /> */}
               <St.IconComment />
             </St.InfoContentIcon>
-            <St.InfoContentCount>
-              {commentCount}
-            </St.InfoContentCount>
+            <St.InfoContentCount>{commentCount}</St.InfoContentCount>
           </St.InfoContent>
         </St.InfoBox>
       </St.PostCardCont>
